+++ EveLib.EveCentral/EveCentralOptions.cs
@@ -10,19 +10,19 @@ namespace eZet.EveLib.Modules {
         ///     Creates a new options object.
         /// </summary>
         public EveCentralOptions() {
-            Items = new List<long>();
-            Regions = new List<long>();
+            Items = new List<int>();
+            Regions = new List<int>();
         }
 
         /// <summary>
         ///     Gets or sets a collection of type IDs.
         /// </summary>
-        public ICollection<long> Items { get; set; }
+        public ICollection<int> Items { get; set; }
 
         /// <summary>
         ///     Gets or sets a collection of region IDs.
         /// </summary>
-        public ICollection<long> Regions { get; set; }
+        public ICollection<int> Regions { get; set; }
 
         /// <summary>
         ///     Gets or sets an age limit for data in the response, specified in hours.
