+++ EveLib.Tests/Element43Legacy_Tests.cs
@@ -7,8 +7,8 @@ namespace eZet.EveLib.Test {
     [TestClass]
     public class Element43Legacy_Tests {
 
-        private const long RegionId = 10000002;
-        private const long TypeId = 34;
+        private const int RegionId = 10000002;
+        private const int TypeId = 34;
         private readonly Element43Legacy _api;
         private readonly Element43Options _validOptions;
